+++ lib/functions/create/create_app_pages.dart
@@ -26,12 +26,36 @@ Future<void> addAppPage(String name) async {
   while (lines.last.isEmpty) {
     lines.removeLast();
   }
+
+  /* 
+  Verificar se as utimas linha segue o padrão esperado.
+  exemplo: 
+        ), // finalizar o getPage 
+      ]; // finalizar o array de pages
+    } // finaliza a classe
+
+    exemplos de fora do padrão:
+    ),];} // tudo na mesma linha
+  */
+  if (lines.last.trim() != '}') {
+    lines.last = lines.last.replaceAll('}', '');
+    lines.add('}');
+  }
+  int index = lines.indexWhere((element) => element.contains('];'));
+  if (lines[index].trim() != '];') {
+    lines[index] = lines[index].replaceAll('];', '');
+    index++;
+    lines.insert(index, '  ];');
+  }
+
   String nameSnakeCase = name.snakeCase;
   String namePascalCase = name.pascalCase;
-  String line = '''    GetPage(name: Routes.${nameSnakeCase.toUpperCase()}, 
+  String line = '''    GetPage(
+      name: Routes.${nameSnakeCase.toUpperCase()}, 
       page:()=> ${namePascalCase}View(), 
-      binding: ${namePascalCase}Binding()),''';
-  lines.insert(lines.length - 2, line);
+      binding: ${namePascalCase}Binding(),
+    ),''';
+  lines.insert(index, line);
   lines.insert(
       0, '''import '../pages/$nameSnakeCase/${nameSnakeCase}_binding.dart';''');
   lines.insert(