+++ CHANGELOG.md
@@ -1,5 +1,18 @@
+
+## [0.9.4]
+- Improve code format
+
+## [0.9.3]
+- Remove duplicate appPages
+
+## [0.9.2]
+- Remove duplicate imports
+
+## [0.9.1]
+- New routes format
+
 ## [0.9.0]
-- auto create route when creating page in getx_pattern
+- Auto create route on page create
 
 ## [0.8.8]
 - Auto install get katekko structure